requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
pytesseract==0.3.10
PyMuPDF==1.23.3
python-docx==0.8.11
pdf2image==1.16.3
Pillow==10.0.0
pydub==0.25.1
werkzeug==2.3.7
gunicorn==21.2.0
